requirements.txt
numpy==1.23.5
scikit-image
scipy==1.9.0
scikit-learn
smplx==0.1.28
tqdm
yacs
numba
opencv-python
tensorboardx
filterpy
cython
chumpy
Pillow
trimesh
pyrender
matplotlib
json_tricks
torchgeometry
einops
joblib
boto3
requests
easydict
pycocotools
plyfile
timm
pyglet
psutil
seaborn
